r/aws_ecs_task_definition: Use AWS SDK for Go v2 serializer for 'container_definitions'.
